add logs files

+++ business_cards_app/server/users/routes/usersRestController.js
@@ -1,6 +1,8 @@
 const express = require("express");
 
+const auth = require("../../auth/authService");
 const { handleError } = require("../../utils/handleErrors");
+const { generateUserPassword } = require("../helpers/bcrypt");
 const normalizeUser = require("../helpers/normalizeUser");
 const {
   registerUser,
@@ -11,6 +13,7 @@ const {
   changeUserBusinessStatus,
   deleteUser,
 } = require("../models/usersAccessDataService");
+const userUpdateValidation = require("../validations/Joi/userUpdateValidation");
 const {
   validateRegistration,
   validateLogin,
@@ -27,6 +30,7 @@ router.post("/", async (req, res) => {
       return handleError(res, 400, `Joi Error: ${error.details[0].message}`);
 
     user = normalizeUser(user);
+    user.password = generateUserPassword(user.password);
     user = await registerUser(user);
     return res.status(201).send(user);
   } catch (error) {
@@ -48,33 +52,53 @@ router.post("/login", async (req, res) => {
   }
 });
 
-router.get("/", async (req, res) => {
+router.get("/", auth, async (req, res) => {
   try {
     const users = await getUsers();
+    const user = req.user;
+    if (!user.isAdmin)
+      return handleError(
+        res,
+        403,
+        "Authorization Error: You must be an admin user to see all users in the database"
+      );
     return res.send(users);
   } catch (error) {
     return handleError(res, error.status || 500, error.message);
   }
 });
 
-router.get("/:id", async (req, res) => {
+router.get("/:id", auth, async (req, res) => {
   try {
     const { id } = req.params;
+    const { _id, isAdmin } = req.user;
     const user = await getUser(id);
+    if (_id !== id && !isAdmin)
+      return handleError(
+        res,
+        403,
+        "Authorization Error: You must be an admin or user owner to see the data"
+      );
     return res.send(user);
   } catch (error) {
     return handleError(res, error.status || 500, error.message);
   }
 });
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
   try {
     const { id } = req.params;
+    const { _id } = req.user;
     let user = req.body;
     const { error } = validateUserUpdate(user);
     if (error)
       return handleError(res, 400, `Joi Error: ${error.details[0].message}`);
-
+    if (_id !== id)
+      return handleError(
+        res,
+        403,
+        "Authorization Error: You must be user owner to update the data"
+      );
     user = normalizeUser(user);
     user = await updateUser(id, user);
     return res.send(user);
@@ -83,19 +107,33 @@ router.put("/:id", async (req, res) => {
   }
 });
 
-router.patch("/:id", async (req, res) => {
+router.patch("/:id", auth, async (req, res) => {
   try {
     const { id } = req.params;
+    const { _id, isAdmin } = req.user;
     const user = await changeUserBusinessStatus(id);
+    if (_id !== id && !isAdmin)
+      return handleError(
+        res,
+        403,
+        "Authorization Error: You must be an admin or user owner to change the data"
+      );
     return res.send(user);
   } catch (error) {
     return handleError(res, error.status || 500, error.message);
   }
 });
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
   try {
     const { id } = req.params;
+    const { _id, isAdmin } = req.user;
+    if (_id !== id && !isAdmin)
+      return handleError(
+        res,
+        403,
+        "Authorization Error: You must be an admin or user owner to delete the data"
+      );
     const user = await deleteUser(id);
     return res.send(user);
   } catch (error) {
